+++ .gitignore
@@ -1,2 +1,2 @@
 saved_plugins/*
-plugin_upload_audit.xlsx
+plugin_upload_audit.xlsx__pycache__/
