+++ requirements.txt
@@ -1,2 +1,3 @@
 pandas==2.3.0
 Requests==2.32.3
+openpyxl==3.1.2
